move all downloading functions to new module
